+++ Changelog.md
@@ -1,6 +1,9 @@
 ## Ignition Gazebo 2.x
 
-### Ignition Gazebo 2.xx.xx (2020-xx-xx)
+### Ignition Gazebo 2.18.0 (2020-05-20)
+
+1. Added a `/world/<world_name>/create_multiple` service that parallels the current `/world/<world_name>/create` service. The `create_multiple` service can handle an `ignition::msgs::EntityFactory_V` message that may contain one or more entities to spawn.
+    * [Pull Request 146](https://github.com/ignitionrobotics/ign-gazebo/pull/146)
 
 ### Ignition Gazebo 2.17.0 (2020-05-13)